Numerous minor improvements in crossover operator, transcription and settings.
In crossover operator:
 - Disabled combining all genes from both parents when neither dominates the other.
 - If both parents contain same allele then allow sometimes setting child allele value to anywhere between range of values in parent alleles.
In HyperNEATTranscriberBain when LEO disabled disable a synapse if its efficacy is 0.
Allow specifying normalisation of sums of disjoint, excess and shared gene values in NEAT genome speciation matching.
Allow specifying minimum and maximum speciation threshold values.

+++ src/com/anji/nn/activationfunction/CosineActivationFunction.java
@@ -46,12 +46,10 @@ public class CosineActivationFunction implements ActivationFunction {
 	}
 
 	/**
-	 * Return <code>input</code> with no transformation.
-	 * 
-	 * @see com.anji.nn.activationfunction.ActivationFunction#apply(double)
+	 * Returns cosine(input).
 	 */
 	public double apply(double input) {
-		return (double) Math.cos(input);
+		return Math.cos(input);
 	}
 
 	/**
